requirements.txt
flask==3.0.0
fpdf2==2.7.9
matplotlib==3.8.2
pandas==2.1.4
plotly==5.18.0
pyyaml==6.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
